backend/requirements.txt
fastapi
uvicorn
pillow
python-multipart
ffmpeg-python
static-ffmpeg
opencv-python
torch
torchvision
matplotlib
numpy
sam2
supervision